# table: Encode/Decode handle faster by reducing escaped variables