<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

Current implementation of `EncodeHandle` and `DecodeHandle` uses `binary.Read` and `binary.Write` which accept `interface`s and make variables escape to heap.

### What is changed and how it works?

Use `binary.BigEndian.PutUint64` and `binary.BigEndian.Uint64` to
prevent variables from escaping to heap.

Source code of the related benchmark can be found in this playground:
https://play.golang.org/p/P7IEQkxhu4f

Benchmark result:

```
BenchmarkEncodeWithBuffer-4     10000000               149 ns/op
BenchmarkEncodeWithBuffer-4     10000000               152 ns/op
BenchmarkEncodeWithBuffer-4     10000000               148 ns/op
BenchmarkDecodeWithBuffer-4     20000000                90.5 ns/op
BenchmarkDecodeWithBuffer-4     20000000                87.8 ns/op
BenchmarkDecodeWithBuffer-4     20000000               126 ns/op
BenchmarkEncode-4               100000000               20.7 ns/op
BenchmarkEncode-4               100000000               16.6 ns/op
BenchmarkEncode-4               100000000               17.0 ns/op
BenchmarkDecode-4               2000000000               0.53 ns/op
BenchmarkDecode-4               2000000000               0.53 ns/op
BenchmarkDecode-4               2000000000               0.52 ns/op

```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes

Side effects


Related changes

Release note