@suzaku 2019-09-02
> @suzaku please add some proper labels.

`type/performance` added.